ISIS-3039: some minor simplifications
