+++ extensions/vw/fullcalendar/wkt-integr/src/main/java/org/apache/isis/extensions/fullcalendar/wkt/fullcalendar/ConfigNew.java
@@ -18,30 +18,38 @@
  */
 package org.apache.isis.extensions.fullcalendar.wkt.fullcalendar;
 
-import com.fasterxml.jackson.annotation.JsonRawValue;
-
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonRawValue;
+
+import lombok.Getter;
+import lombok.Setter;
+
 public class ConfigNew implements Serializable {
 
+    private static final long serialVersionUID = 1L;
+
+    @Getter
     private Header headerToolbar = new Header();
     private List<EventSource> eventSources = new ArrayList<EventSource>();
+
+    @Getter @Setter
     private String themeSystem;
+
+    @Getter @Setter
     private boolean selectable = true;
 
     // events
+    @Setter
     private String eventClick;
+    @Setter
     private String select;
 
-    public Header getHeaderToolbar() {
-        return headerToolbar;
-    }
-
-    public ConfigNew add(EventSource eventSource) {
+    public ConfigNew add(final EventSource eventSource) {
         eventSources.add(eventSource);
         return this;
     }
@@ -50,33 +58,17 @@ public class ConfigNew implements Serializable {
         return Collections.unmodifiableList(eventSources);
     }
 
-    public String getThemeSystem() {
-        return themeSystem;
-    }
-
-    public void setThemeSystem(String themeSystem) {
-        this.themeSystem = themeSystem;
-    }
-
     @JsonRawValue
     public String getEventClick() {
         return eventClick;
     }
 
-    public void setEventClick(String eventClick) {
-        this.eventClick = eventClick;
-    }
 
     @JsonRawValue
     public String getSelect() {
         return select;
     }
 
-    public void setSelect(String select) {
-        this.select = select;
-    }
 
-    public boolean isSelectable() {
-        return selectable;
-    }
+
 }
