--- extensions/vw/fullcalendar/wkt-integr/src/main/java/org/apache/isis/extensions/fullcalendar/wkt/fullcalendar/FullCalendar.events.tpl
@@ -1,13 +1,0 @@
-function(info, successCallback, failureCallback) {
-
-    Wicket.Ajax.ajax({
-        "u": "${url}",
-        "dt": "json",
-        "wr":  false,
-        "ep": {
-            "start": info.start.valueOf(),
-            "end": info.end.valueOf(),
-        },
-        "sh": [function(data, textStatus, jqXHR, attrs) { successCallback(jqXHR) }]
-    });
-}
